CATProject/requirements.txt
Appium-Python-Client==1.2.0
enum34==1.1.10
py==1.10.0
pytest==6.2.4
selenium==3.141.0